Exception handling - delete

+++ src/main/java/com/cursojava/spring/services/UserService.java
@@ -2,9 +2,12 @@ package com.cursojava.spring.services;
 
 import com.cursojava.spring.entities.User;
 import com.cursojava.spring.repositories.UserRepository;
+import com.cursojava.spring.services.exceptions.DatabaseException;
 import com.cursojava.spring.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.Id;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,7 +33,14 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user) {
